Renamed to dma_buffer and guest_memory_access_wrapper files
