+++ vm/devices/user_driver_emulated_mock/src/guest_memory_access_wrapper.rs
@@ -1,15 +1,15 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-//! This module is an extension of the lib.rs file from the crate and houses any structs that require impelemnting unsafe traits.
+//! This module defines and provides an unsafe [`GuestMemoryAccess`] trait implementation for [`GuestMemoryAccessWrapper`].
+//! Refer to to struct comment for more details.
+
+// UNSAFETY: underlying struct needs to implement the unsafe [`GuestMemoryAccess`] trait.
+#![expect(unsafe_code)]
 
 use guestmem::GuestMemory;
 use guestmem::GuestMemoryAccess;
-use parking_lot::Mutex;
 use std::ptr::NonNull;
-use std::sync::Arc;
-use user_driver::memory::MappedDmaTarget;
-use user_driver::memory::PAGE_SIZE;
 
 /// The [`GuestMemoryAccessWrapper`] encapsulates types T that already implement [`GuestMemoryAccess`].
 /// It provides the allow_dma switch regardless of the underlying type T.
@@ -52,57 +52,5 @@ impl<T: GuestMemoryAccess> GuestMemoryAccessWrapper<T> {
     pub fn create_test_guest_memory(mem: T, allow_dma: bool) -> GuestMemory {
         let test_backing = GuestMemoryAccessWrapper { mem, allow_dma };
         GuestMemory::new("test mapper guest memory", test_backing)
-    }
-}
-
-/// A representation of a contiguous slice of memory in a larger [`GuestMemory`]
-pub struct DmaBuffer {
-    mem: GuestMemory,
-    pfns: Vec<u64>,
-    state: Arc<Mutex<Vec<u64>>>,
-}
-
-impl DmaBuffer {
-    /// Creates and returns new [`DmaBuffer`] with the given input parameters
-    pub fn new(mem: GuestMemory, pfns: Vec<u64>, state: Arc<Mutex<Vec<u64>>>) -> Self {
-        Self {
-            mem,
-            pfns,
-            state,
-        }
-    }
-}
-
-impl Drop for DmaBuffer {
-    fn drop(&mut self) {
-        let mut state = self.state.lock();
-        for &pfn in &self.pfns {
-            state[pfn as usize / 64] &= !(1 << (pfn % 64));
-        }
-    }
-}
-
-/// SAFETY: we are handing out a VA and length for valid data, propagating the
-/// guarantee from [`GuestMemory`] (which is known to be in a fully allocated
-/// state because we used `GuestMemory::allocate` to create it).
-unsafe impl MappedDmaTarget for DmaBuffer {
-    fn base(&self) -> *const u8 {
-        self.mem
-            .full_mapping()
-            .unwrap()
-            .0
-            .wrapping_add(self.pfns[0] as usize * PAGE_SIZE)
-    }
-
-    fn len(&self) -> usize {
-        self.pfns.len() * PAGE_SIZE
-    }
-
-    fn pfns(&self) -> &[u64] {
-        &self.pfns
-    }
-
-    fn pfn_bias(&self) -> u64 {
-        0
     }
 }
